:recycle: Refactor: 서비스 interface, impl 분리
