+++ backend/src/resolvers/UserResolver.ts
@@ -10,15 +10,19 @@ import {
 import { PostAnswer } from "../entities/PostAnswer";
 import { PostQuestion } from "../entities/PostQuestion";
 import { User } from "../entities/User";
-import PostService from "../services/PostService";
-import UserService from "../services/UserService";
+import UserService from "../services/User/UserService";
 import Container, { Service } from "typedi";
+import QuestionService from "../services/Question/QuestionService";
+import AnswerService from "../services/Answer/AnswerService";
 
 @Service()
 @Resolver(User)
 export default class UserResolver {
   private readonly userService: UserService = Container.get("UserService");
-  private readonly postService: PostService = Container.get("PostService");
+  private readonly questionService: QuestionService =
+    Container.get("QuestionService");
+  private readonly answerService: AnswerService =
+    Container.get("AnswerService");
 
   @Query(() => User, {
     description: "User의 고유 ID를 통해 유저를 검색",
@@ -46,14 +50,14 @@ export default class UserResolver {
 
   @FieldResolver(() => [PostQuestion], { nullable: "items" })
   async postQuestions(@Root() user: User): Promise<PostQuestion[]> {
-    const questions = await this.postService.findAllQuestionByUserId(user.id);
+    const questions = await this.questionService.findAllByUserId(user.id);
 
     return questions;
   }
 
   @FieldResolver(() => [PostAnswer], { nullable: "items" })
   async postAnswers(@Root() user: User): Promise<PostAnswer[]> {
-    const answers = await this.postService.findAllAnswerByUserId(user.id);
+    const answers = await this.answerService.findAllByUserId(user.id);
 
     return answers;
   }
@@ -84,6 +88,6 @@ export default class UserResolver {
 
   @Query(() => [User], { description: "유저 점수 역순으로 5개 가져오기" })
   async getUsersRank(): Promise<User[]> {
-    return await this.userService.getUsersRank();
+    return await this.userService.getRank();
   }
 }
